Fix examples compilation fail in T3, T5 (#27)
* fix: Stop build when t2 in drawin

* fix: t2 build wifi examples fail

* feat: Supports calling the vendor interface

* fix: Replace the vendor api with the tal api

* fix: all station info list memory is not freed

* feat: Add file system

* fix: Spell error

* fix: Spell error

* feat: add upload speed options and update upload tool configuration

* fix: update WiFiAP and WiFiSTA to include necessary headers and improve IP configuration handling

* fix: ensure setSDA and setSCL methods return true

* feat: Support T3 SPI

* Update package.json

* Update package_cn.json

* fix: update CMakeLists.txt for board configuration and library inclusion

---------

+++ libraries/WiFi/src/WiFiScan.h
@@ -9,7 +9,7 @@ class WiFiScanClass
 public:
     int16_t scanNetworks(bool async = false, bool show_hidden = false, bool passive = false, uint32_t max_ms_per_chan = 300, uint8_t channel = 0, const char * ssid=nullptr, const uint8_t * bssid=nullptr);
     void scanDelete();
-    bool getNetworkInfo(uint8_t networkItem, String &ssid, uint8_t &encryptionType, int32_t &RSSI, uint8_t* &BSSID, int32_t &channel);
+    bool getNetworkInfo(uint32_t networkItem, String &ssid, uint8_t &encryptionType, int32_t &RSSI, uint8_t* &BSSID, int32_t &channel);
 
     String SSID(uint8_t networkItem);
     WF_AP_AUTH_MODE_E encryptionType(uint8_t networkItem);
@@ -17,10 +17,10 @@ public:
     uint8_t * BSSID(uint8_t networkItem);
     String BSSIDstr(uint8_t networkItem);
     int32_t channel(uint8_t networkItem);
-    void * getScanInfoByIndex(int i) { return _getScanInfoByIndex(i); }; 
+    void * getScanInfoByIndex(uint32_t i) { return _getScanInfoByIndex(i); };
 
 protected:
-    void * _getScanInfoByIndex(int i);
+    void * _getScanInfoByIndex(uint32_t i);
 
 private:
     AP_IF_S * ap_info;
